center calculate Points function on another file
